+++ src/index.js
@@ -183,17 +183,32 @@ const PATH_PREFIX = process.env.GITHUB_WORKSPACE;
     }
 
     /**
+     * pagination is not possible see https://github.com/octokit/rest.js/issues/33
+     *
      * @returns {Promise<Record<string, object>>}
      */
     const getReviewers = async () => {
-        // pagination is not possible see https://github.com/octokit/rest.js/issues/33
-        const listReviews = (await octokit.rest.pulls.listReviews({
-            ...repo,
-            pull_number,
-            per_page: 100,
-        }));
+        const route = `GET /repos/${repo.owner}/${repo.repo}/pulls/${pull_number}/reviews`;
+        const options = {per_page: 100};
 
-        const allReviewersData = listReviews.data;
+        const response = await octokit.request(route, options);
+
+        const nextPages = utils.getNextPages(response.headers, route);
+
+        let allReviewersData;
+
+        if(!nextPages) {
+            allReviewersData = response.data;
+        } else {
+            allReviewersData = [
+                response.data,
+                await Promise.all(
+                    nextPages.map(async (page) => {
+                        return (await octokit.request(page, options)).data;
+                    }),
+                ),
+            ].flat(2);
+        }
 
         const latestReviews = {};