+++ CHANGELOG.md
@@ -3,6 +3,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+
+### Fixed
+- [#8](https://github.com/zattoo/reviewers/issues/8) Handle `listReviews` pagination
+
+
 ## 2.0.0 - 29.08.2021
 
 ### Added